moved core data to app delegate

+++ OBSLY/OBMyBooksViewController.swift
@@ -28,28 +28,27 @@ class OBMyBooksViewController: UIViewController, UITableViewDelegate, UITableVie
         setupUI()
         navigationItem.rightBarButtonItem = editButtonItem
         
-//        NSNotificationCenter.defaultCenter().addObserver(self, selector: #selector(OBMyBooksViewController.updateBooksTable), name: "updateBooksTable", object: nil)
-        
     }
+    
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         
-        let managedContext = DataController().managedObjectContext
-        let fetchRequest: NSFetchRequest<Book> = Book.fetchRequest()
-        do {
-            let results =
-                try managedContext.fetch(fetchRequest)
-            books = results
-        } catch let error as NSError {
-            print("Could not fetch \(error), \(error.userInfo)")
+        if #available(iOS 10.0, *) {
+            let managedContext = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
+            let fetchRequest: NSFetchRequest<Book> = Book.fetchRequest()
+            do {
+                let results =
+                    try managedContext.fetch(fetchRequest)
+                books = results
+            } catch let error as NSError {
+                print("Could not fetch \(error), \(error.userInfo)")
+            }
+            booksListTableView.reloadData()
+        } else {
+            print("error: old iOS version")
         }
-        booksListTableView.reloadData()
+        
     }
-    
-//    override func viewWillDisappear(_ animated: Bool) {
-//        super.viewWillDisappear(animated)
-//        NotificationCenter.default.removeObserver(self, name: NSNotification.Name(rawValue: "updateBooksTable"), object: nil)
-//    }
     
 //MARK: UITableViewDataSource
     func numberOfSections(in tableView: UITableView) -> Int {
@@ -65,7 +64,6 @@ class OBMyBooksViewController: UIViewController, UITableViewDelegate, UITableVie
         let index = (indexPath as NSIndexPath).row
         cell.titleLabel.text = "\(index + 1)."
         let book = books[indexPath.row]
-        
         cell.valueLabel.text = book.value(forKey: "bookName") as! String?
         
         // make separator line between cells to fill full width
@@ -82,28 +80,40 @@ class OBMyBooksViewController: UIViewController, UITableViewDelegate, UITableVie
         self.performSegue(withIdentifier: openBookSegueIdentifier, sender: nil)
     }
 
+    // for deleting (editing)
     func tableView(_ tableView: UITableView, commit editingStyle: UITableViewCellEditingStyle, forRowAt indexPath: IndexPath) {
         if editingStyle == .delete {
             // remove the deleted item from the сore date
             
-            let managedContext = DataController().managedObjectContext
-            managedContext.delete(books[indexPath.row] as NSManagedObject)
-            books.remove(at: indexPath.row)
-            do {
-                try managedContext.save()
-            } catch let error as NSError  {
-                print("Could not save \(error), \(error.userInfo)")
+            if #available(iOS 10.0, *) {
+                let managedContext = (UIApplication.shared.delegate as! AppDelegate).persistentContainer.viewContext
+                managedContext.delete(books[indexPath.row] as NSManagedObject)
+                books.remove(at: indexPath.row)
+                do {
+                    try managedContext.save()
+                } catch let error as NSError  {
+                    print("Could not save \(error), \(error.userInfo)")
+                }
+                // Delete the row from the data source
+                tableView.deleteRows(at: [indexPath], with: .fade)
+                // delay to reload the table
+                let when = DispatchTime.now() + 0.1
+                DispatchQueue.main.asyncAfter(deadline: when){
+                    self.booksListTableView.reloadData()
+                }
+            } else {
+                print("error: old iOS version")
             }
-            // Delete the row from the data source
-            tableView.deleteRows(at: [indexPath], with: .fade)
         }
     }
     
+    // for deleting (editing)
     func tableView(_ tableView: UITableView, canEditRowAt indexPath: IndexPath) -> Bool {
         // Return false if you do not want the specified item to be editable.
         return true
     }
     
+    // for deleting (editing)
     override func setEditing(_ editing: Bool, animated: Bool) {
         super.setEditing(editing, animated: animated)
         booksListTableView.setEditing(editing, animated: animated)
@@ -143,5 +153,8 @@ class OBMyBooksViewController: UIViewController, UITableViewDelegate, UITableVie
 
     }
     
+    func reloadTableView(timer: Timer)  {
+        booksListTableView.reloadData()
+    }
     
 }
